+++ components/food/entity.vue
